% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_provenance.R
\name{get_provenance}
\alias{get_provenance}
\title{get_provenance}
\usage{
get_provenance(
  data_product,
  version,
  namespace,
  width = NULL,
  attributes = FALSE,
  endpoint = "http://localhost:8000/api/"
)
}
\arguments{
\item{data_product}{data_product}

\item{version}{version}

\item{namespace}{namespace}

\item{width}{width}

\item{attributes}{attributes}

\item{endpoint}{endpoint}
}
\description{
get_provenance
}
