% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_model.R
\name{initialise_model}
\alias{initialise_model}
\title{initialise_model}
\usage{
initialise_model(rts_params, efoi_params, static_params)
}
\arguments{
\item{rts_params}{rts_params}

\item{efoi_params}{efoi_params}

\item{static_params}{static_params}
}
\description{
initialise_model
}
