% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_SEINRD.R
\name{initialise_SEINRD}
\alias{initialise_SEINRD}
\title{initialise_SEINRD}
\usage{
initialise_SEINRD(rts_params, efoi_params, static_params)
}
\arguments{
\item{rts_params}{rts_params}

\item{efoi_params}{efoi_params}

\item{static_params}{static_params}
}
\description{
initialise_SEINRD
}
