% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEINRD_model.R
\name{SEINRD_model}
\alias{SEINRD_model}
\title{SEINRD_model}
\usage{
SEINRD_model(time, state, parms)
}
\arguments{
\item{time}{time}

\item{state}{state}

\item{parms}{parms}
}
\description{
SEINRD_model
}
