% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRS_model.R
\name{SEIRS_model}
\alias{SEIRS_model}
\title{SEIRS_model}
\usage{
SEIRS_model(
  S,
  E,
  I,
  R,
  timesteps,
  alpha,
  beta,
  inv_gamma,
  inv_omega,
  inv_mu,
  inv_sigma
)
}
\arguments{
\item{S}{Susceptible individuals at time, t = 0}

\item{E}{Exposed individuals at time, t = 0}

\item{I}{Infectious individuals at time, t = 0}

\item{R}{Recovered individuals at time, t = 0}

\item{timesteps}{timesteps}

\item{alpha}{alpha}

\item{beta}{Contact rate}

\item{inv_gamma}{Infectious period}

\item{inv_omega}{Average protected period}

\item{inv_mu}{inverse mu}

\item{inv_sigma}{Latency period}
}
\description{
SEIRS_model
}
