% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEIRS_model.R
\name{SEIRS_model}
\alias{SEIRS_model}
\title{SEIRS_model}
\usage{
SEIRS_model(
  initial.state,
  timesteps,
  years,
  alpha,
  beta,
  inv_gamma,
  inv_omega,
  inv_mu,
  inv_sigma
)
}
\arguments{
\item{initial.state}{Number of individuals in different disease states (S, E, I and R) at time, t = 0}

\item{timesteps}{timesteps}

\item{years}{years}

\item{alpha}{Death rate}

\item{beta}{Contact rate (per day)}

\item{inv_gamma}{Infectious period (days)}

\item{inv_omega}{Average protected period (years)}

\item{inv_mu}{inverse mu (years)}

\item{inv_sigma}{Latency period (days)}
}
\description{
SEIRS_model
}
